+++ src/main/java/it/matjaz/numerus/RomanNumeral.java
@@ -3,7 +3,7 @@
  *
  * This Source Code Form is part of the project Numerus, a roman numerals
  * library for Java. The library and its source code may be found on:
- * https://github.com/MatjazDev/Numerus and http://matjaz.it/numerus/
+ * https://github.com/TheMatjaz/Numerus and http://matjaz.it/numerus/
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -24,9 +24,9 @@ import java.util.regex.Pattern;
  * matches the {@link #CORRECT_ROMAN_SYNTAX_REGEX}.
  * <p>
  * Any string with other characters, different order, too many characters or
- * anyhow incorrect syntax gets refused with {@link NumberFormatException},
- * which may contain some indication of the syntax error in the Exception
- * message.
+ * anyhow incorrect syntax gets refused with
+ * {@link IllegalNumeralSyntaxException}, which may contain some indication of
+ * the syntax error in the Exception message.
  * <p>
  * The structure of a syntactically roman numeral is composed of the following
  * characters <i>in this order</i>:
@@ -93,6 +93,11 @@ public class RomanNumeral implements Serializable, Cloneable, CharSequence {
     private static final String TWO_SAME_FIVE_LIKE_CHARS_REGEX = "(D.*D|L.*L|V.*V)";
 
     /**
+     * String indicating the roman numeral with value zero, 0.
+     */
+    public static final String NULLA = "NULLA";
+
+    /**
      * Serializable class version number.
      * <p>
      * It is used during deserialization to verify that the sender and receiver
@@ -106,29 +111,31 @@ public class RomanNumeral implements Serializable, Cloneable, CharSequence {
      *
      * @see Serializable
      */
-    private static final long serialVersionUID = 20150422L;
+    private static final long serialVersionUID = 20150903L;
 
     /**
-     * Constructs a RomanNumeral initialized to "NULLA".
+     * Constructs a RomanNumeral initialized to NULLA.
      * <p>
      * Contains <b>"NULLA"</b> numerals indicating zero value.
      */
     public RomanNumeral() {
-        this.numeral = "NULLA";
+        this.numeral = NULLA;
     }
 
     /**
      * Constructs a RomanNumeral initialized with the given value.
      * <p>
      * The passed string gets checked for syntax correctness. If the syntax is
-     * illegal, then a {@link NumberFormatException} is thrown.
+     * illegal, then a {@link IllegalNumeralSyntaxException} is thrown.
      * <p>
      * Whitespace charactes in the passed String are removed and the characters
      * are upcased.
      *
      * @param symbols the initial roman numeral to be stored.
+     * @throws IllegalNumeralSyntaxException when string has illegal roman
+     * syntax.
      */
-    public RomanNumeral(String symbols) {
+    public RomanNumeral(String symbols) throws IllegalNumeralSyntaxException {
         this.numeral = cleanUpcaseAndSyntaxCheckString(symbols);
     }
 
@@ -155,21 +162,23 @@ public class RomanNumeral implements Serializable, Cloneable, CharSequence {
      * <code>false</code>.
      */
     public boolean isNulla() {
-        return "NULLA".equals(getNumeral());
+        return NULLA.equals(getNumeral());
     }
 
     /**
      * Setter of the roman numerals String.
      * <p>
      * The passed string gets checked for syntax correctness. If the syntax is
-     * illegal, then a {@link NumberFormatException} is thrown.
+     * illegal, then a {@link IllegalNumeralSyntaxException} is thrown.
      * <p>
      * Whitespace charactes in the passed String are removed and the characters
      * are upcased.
      *
      * @param numeral the new roman numeral to be stored.
+     * @throws IllegalNumeralSyntaxException when the passed string has illegal
+     * roman syntax.
      */
-    public void setNumeral(String numeral) {
+    public void setNumeral(String numeral) throws IllegalNumeralSyntaxException {
         this.numeral = cleanUpcaseAndSyntaxCheckString(numeral);
     }
 
@@ -193,7 +202,7 @@ public class RomanNumeral implements Serializable, Cloneable, CharSequence {
         try {
             new RomanNumeral().cleanUpcaseAndSyntaxCheckString(numeralsToCheck);
             return true;
-        } catch (NumberFormatException ex) {
+        } catch (IllegalNumeralSyntaxException ex) {
             return false;
         }
     }
@@ -202,12 +211,13 @@ public class RomanNumeral implements Serializable, Cloneable, CharSequence {
      * Removes all whitespace characters, upcases the String and verifies the
      * roman syntax.
      * <p>
-     * If the syntax does not match, an exception is thrown.
+     * If the syntax does not match, a {@link IllegalNumeralSyntaxException} is
+     * thrown.
      *
      * @param symbols string to be cleaned, upcased and checked.
      * @return given string without whitespaces and upcased.
      */
-    private String cleanUpcaseAndSyntaxCheckString(String symbols) {
+    private String cleanUpcaseAndSyntaxCheckString(String symbols) throws IllegalNumeralSyntaxException {
         String cleanSymbols = symbols.replaceAll("\\s+", "").toUpperCase();
         throwExceptionIfIllegalRomanSyntax(cleanSymbols);
         return cleanSymbols;
@@ -235,28 +245,37 @@ public class RomanNumeral implements Serializable, Cloneable, CharSequence {
         return matches.toString();
     }
 
-    private void throwExceptionIfIllegalRomanSyntax(String symbols) {
+    /**
+     * Performs a check of the roman syntax of the given string.
+     *
+     * If the syntax is not correct, a {@link IllegalNumeralSyntaxException} is
+     * thrown with some specification about the error. The syntax is indicated
+     * in {@link RomanNumeral#CORRECT_ROMAN_SYNTAX_REGEX}.
+     *
+     * @param symbols
+     */
+    private void throwExceptionIfIllegalRomanSyntax(String symbols) throws IllegalNumeralSyntaxException {
         if (symbols.isEmpty()) {
-            throw new NumberFormatException("Empty roman numeral.");
+            throw new IllegalNumeralSyntaxException("Empty roman numeral.");
         }
         if (symbols.length() >= 20) {
-            throw new NumberFormatException("Impossibly long roman numeral.");
+            throw new IllegalNumeralSyntaxException("Impossibly long roman numeral.");
         }
         if (!symbols.matches(CORRECT_ROMAN_SYNTAX_REGEX)) {
             String illegalChars;
             illegalChars = findAllRegexMatchingSubstrings(symbols, NON_ROMAN_CHARS_REGEX);
             if (!illegalChars.isEmpty()) {
-                throw new NumberFormatException("Non roman characters: " + illegalChars);
+                throw new IllegalNumeralSyntaxException("Non roman characters: " + illegalChars);
             }
             illegalChars = findAllRegexMatchingSubstrings(symbols, FOUR_CONSECUTIVE_TEN_LIKE_CHARS_REGEX);
             if (!illegalChars.isEmpty()) {
-                throw new NumberFormatException("Four consecutive: " + illegalChars);
+                throw new IllegalNumeralSyntaxException("Four consecutive: " + illegalChars);
             }
             illegalChars = findAllRegexMatchingSubstrings(symbols, TWO_SAME_FIVE_LIKE_CHARS_REGEX);
             if (!illegalChars.isEmpty()) {
-                throw new NumberFormatException("Two D, L or V same characters: " + illegalChars);
+                throw new IllegalNumeralSyntaxException("Two D, L or V same characters: " + illegalChars);
             }
-            throw new NumberFormatException("Generic roman numeral syntax error.");
+            throw new IllegalNumeralSyntaxException("Generic roman numeral syntax error.");
         }
     }